#11 #12 #20 Moving files around
Organizing these files into packages in anticipation of more classes
being written.
