+++ src/main/java/edu/southwestern/MMNEAT/MMNEAT.java
@@ -61,7 +61,7 @@ import edu.southwestern.tasks.gridTorus.TorusPredPreyTask;
 import edu.southwestern.tasks.gridTorus.competitive.CompetitiveHomogeneousPredatorsVsPreyTask;
 import edu.southwestern.tasks.gridTorus.cooperative.CooperativePredatorsVsStaticPreyTask;
 import edu.southwestern.tasks.gridTorus.cooperativeAndCompetitive.CompetitiveAndCooperativePredatorsVsPreyTask;
-import edu.southwestern.tasks.gvgai.GVGAISinglePlayerTask;
+import edu.southwestern.tasks.gvgai.player.GVGAISinglePlayerTask;
 import edu.southwestern.tasks.innovationengines.PictureInnovationTask;
 import edu.southwestern.tasks.innovationengines.ShapeInnovationTask;
 import edu.southwestern.tasks.interactive.InteractiveEvolutionTask;
