+++ src/main/java/edu/southwestern/tasks/gvgai/GVGAIUtil.java
@@ -1,0 +1,5 @@
+package edu.southwestern.tasks.gvgai;
+
+public class GVGAIUtil {
+
+}
