+++ src/main/java/edu/southwestern/tasks/gvgai/player/GVGAIReactiveNNPlayer.java
@@ -1,4 +1,4 @@
-package edu.southwestern.tasks.gvgai;
+package edu.southwestern.tasks.gvgai.player;
 
 import java.awt.Point;
 import java.util.ArrayList;
