+++ src/main/java/edu/southwestern/tasks/gvgai/player/GVGAITreeSearchNNPlayer.java
@@ -1,4 +1,4 @@
-package edu.southwestern.tasks.gvgai;
+package edu.southwestern.tasks.gvgai.player;
 
 import java.util.List;
 import java.util.Random;
